+++ src/timeline.js
@@ -8,7 +8,7 @@ import Measure from 'react-measure';
 import interact from 'interactjs';
 import _ from 'lodash';
 import {Button, Popup} from 'semantic-ui-react';
-import {Group, InteractOption, Item, RowLayer} from './index';
+import {Group, InteractOption, Item, RowLayer, getDurationFromPixels} from './index';
 
 import {TestsAreDemoCheat, createTestids} from '@famiprog-foundation/tests-are-demo';
 import {Marker} from './components/Marker';
@@ -41,14 +41,14 @@ import {
 import 'core-js/fn/string/starts-with';
 import SplitPane from 'react-split-pane';
 import 'fixed-data-table-2/dist/fixed-data-table.css';
-// TODO DB this is temporary. Created a copy because if we have used ItemRenderer file the default property for itemRenderer property
-// didn't work and it gets undefined
 import ItemRenderer from './components/ItemRenderer';
 import {SelectionHolder} from './utils/SelectionHolder';
 import {IGanttAction} from './types';
 import {ContextMenu} from './components/ContextMenu/ContextMenu';
 import {GanttContext} from './utils/GanttContext';
 import {SimpleGanttTable} from './components/SimpleGanttTable';
+import moment from 'moment';
+import {SCROLLBAR_SIZE, Scrollbar} from './components/Scrollbar';
 
 const testids = createTestids('Timeline', {
   menuButton: '',
@@ -158,7 +158,7 @@ export default class Timeline extends React.Component {
      * By setting this property the gantt selection mechanism (mentioned above) is disabled, and the selected items are only the ones "dictated" by this property
      * @type { Array.<number | string> }
      */
-    selectedItems: PropTypes.arrayOf(PropTypes.number | PropTypes.string),
+    selectedItems: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.number, PropTypes.string])),
 
     /**
      * The component that is the item (segment) renderer. You can change the default component (i.e. `ItemRenderer`). We
@@ -237,16 +237,36 @@ export default class Timeline extends React.Component {
     rowOddClassName: PropTypes.string,
 
     /**
-     * Start of the displayed interval, as date (numeric/millis or moment object, cf. `useMoment`).
+     * The start of the displayed interval, as date (numeric/millis or moment object, cf. `useMoment`).
+     *
+     * This is only an initial value, because the display interval can change by horizontally srolling the diagram
      * @type {number | object}
      */
     startDate: PropTypes.oneOfType([PropTypes.number, PropTypes.object]).isRequired,
 
     /**
      * End of the displayed interval, as date (numeric/millis or moment object, cf. `useMoment`).
+     *
+     * This is only an initial value, because the display interval can change by horizontally srolling the diagram
      * @type { number | object }
      */
     endDate: PropTypes.oneOfType([PropTypes.number, PropTypes.object]).isRequired,
+
+    /**
+     * The minimum start of the display interval that can be reached by horizontally scrolling the gantt diagram
+     * If not provided, the `startDate` will also be the the minimum. So no scroll will be possible beyond `startDate`
+     *
+     * @type { number | object }
+     */
+    minDate: PropTypes.oneOfType([PropTypes.number, PropTypes.object]),
+
+    /**
+     * The maximum end of the display interval that can be reached by horizontally scrolling the gantt diagram.
+     * If not provided, the `endDate` will also be the the maximum. So no scroll will be possible beyond `endDate`
+     *
+     * @type { number | object }
+     */
+    maxDate: PropTypes.oneOfType([PropTypes.number, PropTypes.object]),
 
     /** If `false`, then when you "talk" dates/times to the Timeline, then you use
      * plain timestamps (i.e. number of millis, e.g. `new Date().valueOf()`). And this everywhere where
@@ -437,7 +457,15 @@ export default class Timeline extends React.Component {
      *
      * @type {(param: IGanttOnContextMenuShowParam) => IGanttAction[]}
      */
-    onContextMenuShow: PropTypes.func
+    onContextMenuShow: PropTypes.func,
+
+    /**
+     * If this handler is provided, it will be called when the table is resized
+     * by dragging the split bar between the table and the gantt.
+     *
+     * @type {(tableWidth: number) => void}
+     */
+    onTableResize: PropTypes.func
   };
 
   static defaultProps = {
@@ -463,6 +491,8 @@ export default class Timeline extends React.Component {
     // in rtl9k:
     // useMoment: true,
     useMoment: false,
+    minDate: undefined,
+    maxDate: undefined,
     selectedItems: undefined,
     snap: undefined,
     timebarFormat: undefined,
@@ -481,7 +511,8 @@ export default class Timeline extends React.Component {
     onDragToCreateStarted: undefined,
     onDragToCreateEnded: undefined,
     onContextMenuShow: undefined,
-    onSelectionChange() {}
+    onSelectionChange() {},
+    onTableResize: undefined
   };
 
   /**
@@ -511,8 +542,9 @@ export default class Timeline extends React.Component {
    */
   static no_op = () => {};
 
-  getInitialTableWidth() {
-    return this.props.table ? this.props.table.props.width : 0;
+  getInitialTableWidth(props) {
+    props = props ? props : this.props;
+    return props.table ? props.table.props.width : 0;
   }
 
   constructor(props) {
@@ -535,7 +567,11 @@ export default class Timeline extends React.Component {
       scrollTop: 0,
       openedContextMenuCoordinates: undefined,
       openedContextMenuRow: undefined,
-      openedContextMenuTime: undefined
+      openedContextMenuTime: undefined,
+      startDate: this.props.startDate,
+      endDate: this.props.endDate,
+      hasHorizontalScrollbar: false,
+      touchPositionX: undefined
     };
 
     // These functions need to be bound because they are passed as parameters.
@@ -563,6 +599,7 @@ export default class Timeline extends React.Component {
     this.updateDimensions = this.updateDimensions.bind(this);
     this.grid_ref_callback = this.grid_ref_callback.bind(this);
     this.table_ref_callback = this.table_ref_callback.bind(this);
+    this.horizontalScroll_ref_callback = this.horizontalScroll_ref_callback.bind(this);
     this.select_ref_callback = this.select_ref_callback.bind(this);
     this.selectionHolder_ref_callback = this.selectionHolder_ref_callback.bind(this);
     this.menuButton_ref_callback = this.menuButton_ref_callback.bind(this);
@@ -570,6 +607,9 @@ export default class Timeline extends React.Component {
     this.mouseMoveFunc = this.mouseMoveFunc.bind(this);
     this.mouseDownFunc = this.mouseDownFunc.bind(this);
     this.mouseUpFunc = this.mouseUpFunc.bind(this);
+    this.onTouchStart = this.onTouchStart.bind(this);
+    this.onTouchMove = this.onTouchMove.bind(this);
+    this.onTouchEnd = this.onTouchEnd.bind(this);
     this.getCursor = this.getCursor.bind(this);
     this.setVerticalGridLines = this.setVerticalGridLines.bind(this);
     this.handleScrollTable = this.handleScrollTable.bind(this);
@@ -589,13 +629,23 @@ export default class Timeline extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    this.setTimeMap(
-      nextProps.items,
-      convertDateToMoment(nextProps.startDate, nextProps.useMoment),
-      convertDateToMoment(nextProps.endDate, nextProps.useMoment),
-      nextProps.useMoment
-    );
+    const tableWidth = this.getInitialTableWidth(nextProps);
+    if (tableWidth != this.state.tableWidth) {
+      this.setState({tableWidth});
+    }
+
+    if (this.props.startDate != nextProps.startDate || this.props.endDate != nextProps.endDate) {
+      this.setState({startDate: this.props.startDate, endDate: this.props.endDate});
+    } else {
+      this.setTimeMap(
+        nextProps.items,
+        convertDateToMoment(nextProps.startDate, nextProps.useMoment),
+        convertDateToMoment(nextProps.endDate, nextProps.useMoment),
+        nextProps.useMoment
+      );
+    }
     this.fillInTimelineWithEmptyRows(nextProps.groups);
+
     // @TODO
     // investigate if we need this, only added to refresh the grid
     // when double click -> add an item
@@ -620,6 +670,16 @@ export default class Timeline extends React.Component {
       const canResize = Timeline.isBitSet(Timeline.TIMELINE_MODES.RESIZE, timelineMode);
       this.setUpDragging(canSelect, canDrag, canResize);
     }
+
+    if (prevState.startDate != this.state.startDate || prevState.endDate != this.state.endDate) {
+      this.setTimeMap(
+        this.props.items,
+        convertDateToMoment(this.state.startDate, this.props.useMoment),
+        convertDateToMoment(this.state.endDate, this.props.useMoment),
+        this.props.useMoment
+      );
+      this.refreshGrid();
+    }
   }
 
   /**
@@ -638,7 +698,7 @@ export default class Timeline extends React.Component {
    * @return {moment}
    */
   getStartDate() {
-    return convertDateToMoment(this.props.startDate, this.props.useMoment);
+    return convertDateToMoment(this.state.startDate, this.props.useMoment);
   }
 
   /**
@@ -647,7 +707,33 @@ export default class Timeline extends React.Component {
    * @return {moment}
    */
   getEndDate() {
-    return convertDateToMoment(this.props.endDate, this.props.useMoment);
+    return convertDateToMoment(this.state.endDate, this.props.useMoment);
+  }
+
+  /**
+   * Min of the displayed interval (as moment object)
+   *
+   * @returns {moment}
+   */
+  getMinDate() {
+    if (this.props.minDate) {
+      return convertDateToMoment(this.props.minDate, this.props.useMoment);
+    } else {
+      return this.getStartDate();
+    }
+  }
+
+  /**
+   * Max of the displayed interval (as moment object)
+   *
+   * @returns {moment}
+   */
+  getMaxDate() {
+    if (this.props.maxDate) {
+      return convertDateToMoment(this.props.maxDate, this.props.useMoment);
+    } else {
+      return this.getEndDate();
+    }
   }
 
   /**
@@ -778,8 +864,18 @@ export default class Timeline extends React.Component {
         this.itemRowMap[item.key] = rowInt;
         this.rowItemMap[rowInt].push(item);
       });
+    });
+
+    let maxVisibleItems = _.filter(items, i => {
+      return (
+        this.getEndFromItem(i, useMoment) > this.getMinDate() && this.getStartFromItem(i, useMoment) < this.getMaxDate()
+      );
+    });
+    let maxVisibleItemsRows = _.groupBy(maxVisibleItems, 'row');
+    _.forEach(maxVisibleItemsRows, (maxVisibleItems, row) => {
+      const rowInt = parseInt(row);
       this.rowHeightCache[rowInt] = getMaxOverlappingItems(
-        visibleItems,
+        maxVisibleItems,
         this.getStartFromItem,
         this.getEndFromItem,
         useMoment
@@ -812,6 +908,7 @@ export default class Timeline extends React.Component {
       totalItemsHeight += (that.rowHeightCache[group.id] || 1) * that.props.itemHeight;
     });
     let heightToFillIn = this._grid.props.height - totalItemsHeight;
+
     let fillInGroups = [];
 
     let groupId = groups.length;
@@ -939,6 +1036,10 @@ export default class Timeline extends React.Component {
   // because the documnetation fails to be generated
   onDragStartSelect(clientX, clientY) {
     const nearestRowObject = getNearestRowObject(clientX, clientY);
+    if (nearestRowObject == null) {
+      // this happens when you start dragging from e.g. timebar
+      return;
+    }
     const startY = adjustRowTopPositionToViewport(nearestRowObject, nearestRowObject.getBoundingClientRect().y);
     // Add 2 to startY because on some occasions/browsers, when using document.elementsFromPonint(), it will return the wrong row if startY is used.
     // Adding 2 to it ensures that the point isn't shared with other row.
@@ -974,6 +1075,10 @@ export default class Timeline extends React.Component {
     const magicalConstant = 2;
     const {startX, startY} = this._selectBox;
     const startRowObject = getNearestRowObject(startX, startY);
+    if (startRowObject == null) {
+      // this happens when you start dragging from e.g. timebar
+      return;
+    }
     const startXRowObject = startRowObject.getBoundingClientRect().x + 1;
     // select only row without group
     if (startXElement < startXRowObject) {
@@ -1388,7 +1493,15 @@ export default class Timeline extends React.Component {
   }
 
   _handleItemRowEvent = (e, itemCallback, rowCallback) => {
-    e.preventDefault();
+    // First this handler used to be called only on click (no e.preventDefault() was called back then)
+    // After that, it was modified to be called also on contextMenu and doubleClick, then it was added this e.preventDefault()
+    // That's why I suspect that it targets only the contextMenu events in order to prevent the browser default context menu to open
+    // But because this is a legacy code and we don't understand why the exact scope of this preventDefault(),
+    // We removed it only for the case that we are interested in i.e. 'mouseDown'
+    if (e.type != 'mousedown') {
+      e.preventDefault();
+    }
+
     // Skip click handler if selecting with selection box
     if (this.selecting) {
       return;
@@ -1564,7 +1677,7 @@ export default class Timeline extends React.Component {
     var tableRowHeight = this.rowHeight({index});
     let group = _.find(this.state.groups, g => g.id == index);
     if (group.rowHeight && group.key.startsWith(EMPTY_GROUP_KEY)) {
-      tableRowHeight = Math.round(tableRowHeight) - 2;
+      tableRowHeight = Math.round(tableRowHeight) + (this.state.hasHorizontalScrollbar ? SCROLLBAR_SIZE : 0) - 2;
     }
     return tableRowHeight;
   }
@@ -1580,6 +1693,10 @@ export default class Timeline extends React.Component {
 
   table_ref_callback(reactComponent) {
     this._table = reactComponent;
+  }
+
+  horizontalScroll_ref_callback(reactComponent) {
+    this._horizontalScroll = reactComponent;
   }
 
   /**
@@ -1689,6 +1806,51 @@ export default class Timeline extends React.Component {
   }
 
   /**
+   * Toghether with `onTouchMove` and `onTouchEnd` implements the horizontal scroll by dragging the gantt diagram on mobile devices
+   *
+   * @param {*} e
+   */
+  onTouchStart(e) {
+    // We need only to threat touch events on TimelineBody and highligted intervals surface,
+    // but because TimelineBody is based on a third party component adding the handlers
+    // directly on these components was not possible easily (It needed to extend this Grid adding by adding a wrapper div on which to add the touch handlers)
+    // so we added the handler on the parent component and check inside of it to exclude other children like the scrollbar, or the timebar
+    if (
+      e.target.classList.contains('rct9k-horizontal-scrollbar-outter') ||
+      e.target.classList.contains('rct9k-timebar-item')
+    ) {
+      return;
+    }
+
+    const touch = e.touches[0];
+    this.setState({touchPositionX: touch.clientX});
+  }
+
+  /**
+   * Toghether with `onTouchStart` and `onTouchStart` implements the horizontal scroll by dragging the gantt diagram on mobile devices
+   *
+   * @param {*} e
+   */
+  onTouchMove(e) {
+    if (this.state.touchPositionX != undefined) {
+      const touch = e.touches[0];
+      const deltaX = touch.clientX - this.state.touchPositionX;
+      this.setState({touchPositionX: touch.clientX});
+      this._scrollbar.scrollwithDelta(
+        getDurationFromPixels(deltaX, this.getStartDate(), this.getEndDate(), this.getTimelineWidth()).asMilliseconds()
+      );
+    }
+  }
+
+  /**
+   * Toghether with `onTouchStart` and `onTouchMove` implements the horizontal scroll by dragging the gantt diagram on mobile devices
+   *
+   */
+  onTouchEnd() {
+    this.setState({touchPositionX: undefined});
+  }
+
+  /**
    * Setter for verticalGridLines (that will be passed to `BackgroundLayer`).
    * @param { object } verticalGridLines
    */
@@ -1708,6 +1870,9 @@ export default class Timeline extends React.Component {
 
   handleDrag(width) {
     this.setState({tableWidth: width});
+    if (this.props.onTableResize) {
+      this.props.onTableResize(width);
+    }
     return true;
   }
 
@@ -1816,6 +1981,16 @@ export default class Timeline extends React.Component {
     );
   }
 
+  onHorizontalScroll(scrollPosition) {
+    const displayIntervalInMiliseconds = this.getEndDate().diff(this.getStartDate(), 'milliseconds');
+    this.setState({startDate: this.props.useMoment ? moment(scrollPosition) : scrollPosition});
+    this.setState({
+      endDate: this.props.useMoment
+        ? moment(scrollPosition + displayIntervalInMiliseconds)
+        : scrollPosition + displayIntervalInMiliseconds
+    });
+  }
+
   renderGanttPart({bodyHeight, timebarHeight}) {
     const {
       showCursorTime,
@@ -1874,7 +2049,10 @@ export default class Timeline extends React.Component {
                       this.setState({dragCancel: true});
                       this._selectBox.end();
                     }
-                  }}>
+                  }}
+                  onTouchStart={this.onTouchStart}
+                  onTouchMove={this.onTouchMove}
+                  onTouchEnd={this.onTouchEnd}>
                   <SelectBox
                     ref={this.select_ref_callback}
                     className={this.getDragToCreateMode() ? 'rct9k-selector-outer-add' : ''}
@@ -1892,7 +2070,7 @@ export default class Timeline extends React.Component {
                   {markers.map(m => (
                     <Marker
                       key={m.key}
-                      height={this.state.screenHeight}
+                      height={this.state.screenHeight - (this.state.hasHorizontalScrollbar ? SCROLLBAR_SIZE : 0)}
                       top={0}
                       date={0}
                       shouldUpdate={true}
@@ -1905,7 +2083,7 @@ export default class Timeline extends React.Component {
                   <TimelineBody
                     width={this.state.gridWidth}
                     columnWidth={() => this.state.gridWidth}
-                    height={bodyHeight}
+                    height={bodyHeight - (this.state.hasHorizontalScrollbar ? SCROLLBAR_SIZE : 0)}
                     rowHeight={this.rowHeight}
                     rowCount={this.state.groups.length}
                     columnCount={1}
@@ -1915,6 +2093,19 @@ export default class Timeline extends React.Component {
                     forceRedrawFunc={forceRedrawFunc}
                     onScroll={this.handleScrollGantt}
                   />
+                  <Scrollbar
+                    minScrollPosition={this.getMinDate().valueOf()}
+                    maxScrollPosition={this.getMaxDate().valueOf()}
+                    scrollPosition={this.getStartDate().valueOf()}
+                    pageSize={this.getEndDate().valueOf() - this.getStartDate().valueOf()}
+                    hasArrows={true}
+                    onScroll={scrollPosition => {
+                      this.onHorizontalScroll(scrollPosition);
+                    }}
+                    onVisibilityChange={isScrollbarVisible =>
+                      this.setState({hasHorizontalScrollbar: isScrollbarVisible})
+                    }
+                    ref={node => (this._scrollbar = node)}></Scrollbar>
                   {this.renderContextMenu()}
                   {backgroundLayer &&
                     React.cloneElement(backgroundLayer, {
@@ -1922,7 +2113,7 @@ export default class Timeline extends React.Component {
                       endDateTimeline: this.getEndDate(),
                       width: this.state.gridWidth,
                       leftOffset: 0,
-                      height: bodyHeight,
+                      height: bodyHeight - (this.state.hasHorizontalScrollbar ? SCROLLBAR_SIZE : 0),
                       topOffset: timebarHeight,
                       verticalGridLines: this.state.verticalGridLines
                     })}
@@ -2019,7 +2210,7 @@ export default class Timeline extends React.Component {
                   <SplitPane
                     split="vertical"
                     style={{height: this.state.screenHeight}}
-                    defaultSize={this.props.table ? this.getInitialTableWidth() : 0}
+                    size={this.state.tableWidth}
                     onChange={this.handleDrag}
                     ref={this.splitPane_ref_callback}>
                     <GanttContext.Provider value={{rows: this.props.groups, configureTable: this.configureTable}}>